+++ cogs/dota/const.py
@@ -1,6 +1,6 @@
 STEAM_CDN_URL = 'https://cdn.cloudflare.steamstatic.com'
 ODOTA_API_URL = 'https://api.opendota.com/api'
-MANGO_ICON = "https://static.wikia.nocookie.net/dota2_gamepedia/images/3/3d/Greater_Mango_icon.png"
+DISCONNECT_ICON = "https://i.imgur.com/9n8oSge.png"
 MISSING_INFO = ''
 
 dota_player_colour_map = {
