+++ cogs/dotafeed.py
@@ -18,7 +18,7 @@ from .dota.models import PostMatchPlayerData, ActiveMatch, OpendotaRequestMatch
 from .utils.checks import is_guild_owner, is_trustee
 from .utils.context import Context, GuildContext
 from .utils.fpc import FPCBase, TwitchAccCheckCog
-from .utils.var import Clr, Ems, MP, Cid
+from .utils.var import Clr, Ems, MP, Cid, Sid
 
 if TYPE_CHECKING:
     from .utils.bot import AluBot
@@ -497,6 +497,7 @@ class DotaFeedTools(commands.Cog, FPCBase, name='Dota 2'):
         steam='either steamid in any of 64/32/3/2 versions, friend_id or just steam profile link',
         twitch='If you proved twitch handle for "name" then press `True` otherwise `False`',
     )
+    @app_commands.guilds(Sid.wink)
     async def slh_dota_database_add(self, ntr: discord.Interaction, name: str, steam: str, twitch: bool):
         """Slash copy of ext_dota_database_list below"""
         ctx = await Context.from_interaction(ntr)
